requirements.txt
django~=3.2.12
djangorestframework
psycopg2-binary
python-dotenv
gunicorn